# backend/main.go
package main

import (
	docs "CS411-Team-Project/backend/docs"
	"CS411-Team-Project/backend/handlers"

	"github.com/gin-gonic/gin"
	swaggerfiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
)

// @title       CS411 API
// @version     1.0
// @description API for the team's project.

// @BasePath /api/v1

func CORSMiddleware() gin.HandlerFunc {
	return func(c *gin.Context) {

		c.Header("Access-Control-Allow-Origin", "*")
		c.Header("Access-Control-Allow-Credentials", "true")
		c.Header("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
		c.Header("Access-Control-Allow-Methods", "POST,HEAD,PATCH, OPTIONS, GET, PUT")

		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(204)
			return
		}

		c.Next()
	}
}

func main() {
	router := gin.Default()
	router.Use(CORSMiddleware())
	docs.SwaggerInfo.BasePath = "/api/v1"

	v1 := router.Group("/api/v1")
	{
		v1.GET("/recipes", handlers.SearchRecipes)
		v1.GET("/recipe/:id", handlers.GetRecipeInfo)
		v1.POST("/save/:userID/:recipeID", handlers.RecipesSave)
		v1.GET("/getrecipes/:userID", handlers.GetSavedRecipes)
		v1.POST("/save/remove/:userID/:recipeID", handlers.RemoveRecipesSaved)

	}
	router.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerfiles.Handler))
	router.Run(":8080")
}